+++ test/ReactViews/Custom/Chart/FeatureInfoPanelChartSpec.tsx
@@ -1,12 +1,14 @@
-import { ReactTestRenderer, act, create } from "react-test-renderer";
+import {
+  render,
+  screen,
+  waitForElementToBeRemoved
+} from "@testing-library/react";
 import { ThemeProvider } from "styled-components";
-import runLater from "../../../../lib/Core/runLater";
 import CsvCatalogItem from "../../../../lib/Models/Catalog/CatalogItems/CsvCatalogItem";
 import CommonStrata from "../../../../lib/Models/Definition/CommonStrata";
 import TerriaFeature from "../../../../lib/Models/Feature/Feature";
 import Terria from "../../../../lib/Models/Terria";
 import ViewState from "../../../../lib/ReactViewModels/ViewState";
-import { SpecChart } from "../../../../lib/ReactViews/Custom/Chart/FeatureInfoPanelChart";
 import { ChartAttributes } from "../../../../lib/ReactViews/Custom/ChartCustomComponent";
 import CsvChartCustomComponent from "../../../../lib/ReactViews/Custom/CsvChartCustomComponent";
 import CustomComponent, {
@@ -60,77 +62,62 @@ describe("FeatureInfoPanelChart", function () {
   describe("yColumn prop", function () {
     describe("when specified", function () {
       it("renders the correct y-column", async function () {
-        const renderer = renderChart(
+        renderChart(
           `<chart src="test/csv_nongeo/x_height.csv" y-column="plant height"></chart>`,
           context
         );
-        await runLater(() => {}); // yield so that the useEffect in FeatureInfoPanelChart gets a chance to load the chart
-        // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-        const chart = renderer?.root.findByType(SpecChart)!;
-        expect(chart).toBeDefined();
-        expect(
-          chart.findAllByProps({ children: "Plant Height x x" }).length
-        ).toBeGreaterThan(0);
+        await waitForElementToBeRemoved(() =>
+          screen.queryByText("chart.noData")
+        );
+
+        expect(screen.getByText("Plant Height x x")).toBeVisible();
       });
 
       it("renders nothing if the specified y-column does not exist", async function () {
-        const renderer = renderChart(
+        renderChart(
           `<chart src="test/csv_nongeo/x_height.csv" y-column="foo-does-not-exist"></chart>`,
           context
         );
-        await runLater(() => {}); // yield so that the useEffect in FeatureInfoPanelChart gets a chance to load the chart
-        expect(() => renderer?.root.findByType(SpecChart)).toThrow();
+
+        expect(screen.queryByText("chart.noData")).toBeVisible();
       });
     });
 
     describe("otherwise", function () {
       it("renders the first line type chart item", async function () {
-        const renderer = renderChart(
+        renderChart(
           `<chart src="test/csv_nongeo/x_height.csv"></chart>`,
           context
         );
-        await runLater(() => {}); // yield so that the useEffect in FeatureInfoPanelChart gets a chance to load the chart
-        // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-        const chart = renderer?.root.findByType(SpecChart)!;
-        expect(chart).toBeDefined();
-        expect(
-          chart.findAllByProps({ children: "Z x x" }).length
-        ).toBeGreaterThan(0);
+
+        await waitForElementToBeRemoved(() =>
+          screen.queryByText("chart.noData")
+        );
+
+        expect(screen.getByText("Z x x")).toBeVisible();
       });
     });
   });
 
   it("can show a custom X axis label", async function () {
-    const renderer = renderChart(
+    renderChart(
       `<chart src="test/csv_nongeo/x_height.csv" preview-x-label="life-time"></chart>`,
       context
     );
-    await runLater(() => {}); // yield so that the useEffect in FeatureInfoPanelChart gets a chance to load the chart
-    // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-    const chart = renderer?.root.findByType(SpecChart)!;
-    expect(chart).toBeDefined();
-    expect(
-      chart.findAllByProps({ children: "life-time" }).length
-    ).toBeGreaterThan(0);
+
+    await waitForElementToBeRemoved(() => screen.queryByText("chart.noData"));
+
+    expect(screen.getAllByText("life-time")[1]).toBeVisible();
   });
 });
 
 /**
  * Render the given chart markup and return the test renderer.
  */
-function renderChart(
-  chartMarkup: string,
-  context: ProcessNodeContext
-): ReactTestRenderer | undefined {
+function renderChart(chartMarkup: string, context: ProcessNodeContext): void {
   const chartElements = parseCustomHtmlToReact(chartMarkup, context, false, {
     ADD_TAGS: ["chart"],
     ADD_ATTR: ChartAttributes
   });
-  let renderer;
-  act(() => {
-    renderer = create(
-      <ThemeProvider theme={terriaTheme}>{chartElements}</ThemeProvider>
-    );
-  });
-  return renderer;
+  render(<ThemeProvider theme={terriaTheme}>{chartElements}</ThemeProvider>);
 }