+++ CHANGES.md
@@ -15,6 +15,7 @@
 - Pass explicit refs to css transition component to avoid findDomNode usage. #7590
 - TSify FadeIn component. #7590
 - Convert `ViewingControls` to a functional component. #7574
+- Convert `Chart` component of `FeatureInfoPanelChart` to functional component.
 - [The next improvement]
 
 #### 8.9.2 - 2025-03-31